Match current spec (#111)
We match the current spec as of June 28, 2024. There are few user-visible changes, but the implementation of the underlying semantics has changed considerably. Instead of working with digit strings, we work with the concepts of cohort and quantum. We implement cohorts as rational numbers. This warrants bumping the major version of the package.